Added lodash for null checks to fix user id fetching issues
This commit introduces lodash checks in various components to ensure that user id is not null before making any API calls. This is necessary to avoid failed requests when user id does not exist. It also ensures our application behaves as expected when user data is not available such as in public routing. This change was also propagated to other parts of the codebase where a similar check was required. Lastly, an update was made to the @ethang/hooks and eslint-config-ethang versions.