Refactor data fetching strategy to use cache busting
Replaced the previous data fetching and invalidating strategy with the `cacheBust` and the Fetch API. The previous strategy was using `react-query` which had performance issues related to suspending and re-fetching data. By adopting the `cacheBust` approach, we can better control when and how data is fetched and invalidated, improving performance and data coherence throughout the application. Additionally, created a new `useFetch` hook to incorporate this new strategy into existing components.