Replace loading spinner with Skeleton loading component
Replaced the Loading Spinner with a Skeleton loading component to improve perceived performance and user experience. The Skeleton component provides a placeholder while the content is still loading, maintaining the layout structure and preparing the user for the content appearance.

The change is applied in 'loading-icon.tsx', 'card-list.tsx', 'list-card.tsx', 'page.tsx' and 'manage/page.tsx' files where asynchronous data fetches occur. The Skeleton component is customized to fit the size of the expected content.

Additionally, old CSS import of the loading spinner is removed from 'layout.tsx' and Suspense from 'react' is utilized for delaying the render of components until the content gets loaded.