Remove all the actions related to managing the Learning List
The actions removed include those handling the addition, removal and updating of materials in the Learning List. Furthermore, functions managing the favorite, completion and order status of materials in the list have also been removed. These changes were made because the Learning List functionality is undergoing a refactor for improved maintainability and cleaner architecture. Updated functionalities will be implemented in the future updates.