Refactor import statements and modules version update
The import statements of different modules across various files were cleaned up to separate the imports into two categories - type and normal imports. This change was implemented to enhance the readability and organization of the code. Also, dependencies in package.json and the lock file were updated to their latest versions for improved performance and security.